+++ docs/release-notes/change-log.md
@@ -9,6 +9,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### alpha service deployments
+
+* Fix fuzzy matching when endpoint require auth headers
+* Fix panic in "serve" when trying to delete a non-existing deployment
+
 ## v1.1.21
 
 ### Changed